requirements.txt
pandas
requests
beautifulsoup4
google-generativeai
openpyxl